Clear linger if the request is for non-DDS so that it's torn down.
If there is a better network (say Wi-Fi), it replaces the cellular
network avaialble and the network starts lingering. Even if DDS
switch is triggered at this point, the cellular network which is on
non-DDS now continues to linger because unneeded returns false for
a lingering network. Fix is to clear the lingering and reap the
non-DDS networks immediately.

CRs-Fixed: 2191761
CRs-Fixed: 2201309
Change-Id: I5b1e9aeb4374d260949e0fd649700f632d08e987